fix: delete 3d bg
