+++ src/components/register.js
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useAuth } from '../contexts/AuthContext';
 import { notify } from '../ui/toast';
-import NetworkBackground3D from './NetworkBackground3D';
+import NetworkBackground3D from './NetworkBackground2D';
 
 const Register = () => {
   const { register } = useAuth();
